+++ packages/drivers/spanner/test/Crud.test.ts
@@ -8,6 +8,7 @@ import {
 } from '@proteinjs/db';
 import { SpannerDriver } from '../src/SpannerDriver';
 import { getDropTestTable } from './dropTestTable';
+import { TransactionContext } from '@proteinjs/db-transaction-context';
 
 const spannerDriverCrudTests = new SpannerDriver(
   {
@@ -18,7 +19,10 @@ const spannerDriverCrudTests = new SpannerDriver(
   getCrudTestTable
 );
 
-describe('CRUD Tests', crudTests(spannerDriverCrudTests, getDropTestTable(spannerDriverCrudTests)));
+describe(
+  'CRUD Tests',
+  crudTests(spannerDriverCrudTests, new TransactionContext(), getDropTestTable(spannerDriverCrudTests))
+);
 
 const spannerDriverColumnTypesTests = new SpannerDriver(
   {
@@ -31,7 +35,11 @@ const spannerDriverColumnTypesTests = new SpannerDriver(
 
 describe(
   'Column Type Tests',
-  columnTypeTests(spannerDriverColumnTypesTests, getDropTestTable(spannerDriverColumnTypesTests))
+  columnTypeTests(
+    spannerDriverColumnTypesTests,
+    new TransactionContext(),
+    getDropTestTable(spannerDriverColumnTypesTests)
+  )
 );
 
 const spannerDriverDynamicRefColTests = new SpannerDriver(